requirements.txt
Django==3.2.15
django-filter==21.1
djangorestframework==3.14.0
psycopg2==2.9.3








